pyproject.toml
[project]
name = "rootdata-mcp-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi-mcp>=0.3.3",
    "fastmcp>=2.2.5",
    "httpx>=0.28.1",
    "openai>=1.76.0",
    "openai-agents>=0.0.13",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.7",
    "sse-starlette>=2.3.3",
    "uvicorn>=0.34.2",
    "mcp[cli]>=1.6.0",
    "fastapi>=0.115.12",
    "starlette>=0.46.2",
]
